[IMP] website_sale: split python files by classes
