+++ addons/website_sale/models/__init__.py
@@ -19,6 +19,7 @@ from . import res_config_settings
 from . import res_country
 from . import res_partner
 from . import sale_order
+from . import sale_order_line
 from . import website
 from . import website_page
 from . import website_snippet_filter
